requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Logging
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Optional (descomente se precisar)
# redis==5.0.1
# slowapi==0.1.9
# prometheus-fastapi-instrumentator==6.1.0
# sentry-sdk==1.40.0
